Convert the vector data to float when copying them to CARTA::Point

+++ src/Util/Message.cc
@@ -118,15 +118,15 @@ CARTA::Point Message::Point(float x, float y) {
 
 CARTA::Point Message::Point(const casacore::Vector<casacore::Double>& input, int x_index, int y_index) {
     CARTA::Point point;
-    point.set_x(input(x_index));
-    point.set_y(input(y_index));
+    point.set_x((float)input(x_index));
+    point.set_y((float)input(y_index));
     return point;
 }
 
 CARTA::Point Message::Point(const std::vector<casacore::Quantity>& input, int x_index, int y_index) {
     CARTA::Point point;
-    point.set_x(input[x_index].getValue());
-    point.set_y(input[y_index].getValue());
+    point.set_x((float)input[x_index].getValue());
+    point.set_y((float)input[y_index].getValue());
     return point;
 }
